+++ common/src/main/java/net/spell_engine/api/spell/SpellInfo.java
@@ -1,0 +1,5 @@
+package net.spell_engine.api.spell;
+
+import net.minecraft.util.Identifier;
+
+public record SpellInfo(Spell spell, Identifier id) {  }